[ts] workaround inability to properly used inferred type:
`fillBatchesCommands` lacks mapping and that results in not so accurate type of key being inffered from `BitrixCommands` as `number`. That's strange, since function does not guarantee it to be always a number.

Besides, later due to that `number` in index results can't be properly accessed, since `Object.keys` will map keys to strings, while results will have indexed signature for numbers only.