[ts] make `get` method work with create and update params:
It feels strange to use `get` for such tasks, but Bitrix works in such a way that a payload of "add" (create), get and update methods are the same except internals of the `result`.

We can move creation and update into standalone `post` method, but it will be exactly the same as current `get`, which makes it questionable.

Hopefully, it future we could use more advanced types mapping and resolution and use one only generic `call` method.