[ts] move rate limiting responsibility from the client methods to the client itself:
It was an original approach, however it had to be scrapped since `Parameters<F>` were losing generics resolutions on `call` and `batch` functions (seems to be TS limitation).

I've found a more generic approach which made it possible to handle rate limiting in the client itself. That makes `call` and `batch` more pure and thus easier to test.